Fix index exports
